Changed to a better event system
The old event system could not support the listener function being a class
member. The new system uses lambdas to make class member listeners
possible, however due to std::function not implementing any form of
operator== a string id has to be assigned to every listener so that
listeners can be remooved when not needed (E.G.: A button on a main menu
should be removed when starting the game to that a player can't click an
exit button during play).