Added support for fonts based on single spritesheets
Fonts should consist of a single image devided up into squares of equal
size each placed in one horizontal line. This can be loaded as an image
and passed to a font constructor as the first argument. The next argument
should be a string containg the letter that each image represents in the
same order as the image. The final argument is the size of each image.